sandbox/requirements.txt
dataclasses==0.8
PyYAML==6.0
